compat: remove wrong backport of kfree_rcu
Backporting kfree_rcu() in compat does not work as
struct rcu_head->func is always interpreted as a pointer in older
kernel versions, but when using kfree_rcu() it is set to an offset.
rcu_head->func is used in some functions in the kernel not backported
by compat-wireless.

In older kernel versions in __rcu_reclaim just the else path is available.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>