compat: use a generic install mechanism for bin/get-compat-kernels
The script bin/get-compat-kernels used to behave differently
depending on the Linux distribution used. Remove this junk
given that even on Debian based systems there are "conflicts"
being run into when installing all the kernel headers: a higher
version of libc is being asked for newer kernels. Dump this
stuff and instead use the same mechanism to install the
kernel headers for all Linux distributions.

This uses ar and tar. If ar is not available warn the user
that binutils is required.

Signed-off-by: Luis R. Rodriguez <mcgrof@frijolero.org>